directly pass BEASTHOST to feedclient instead of resolving it
at least now the feedclient is able to resolve a hostnames passed as the
data source. thus it's better to not pre-lookup the IP as the IP of a
hostname can change.
this is probably being caught by the healthcheck but it's better to just
let the feedclient deal with it if it's able to